+++ test/test-baseN.cpp
@@ -26,7 +26,7 @@ TEST(baseN, isValid)
 }
 TEST(baseN, sizeEncoded)
 {
-    std::vector<std::pair<uint64_t, std::string>> tests = {
+    std::vector<std::pair<size_t, std::string>> tests = {
         {6, "12341234"},
         {5, "00000000"},
     };
@@ -35,7 +35,7 @@ TEST(baseN, sizeEncoded)
 }
 TEST(baseN, sizeDecoded)
 {
-    std::vector<std::pair<uint64_t, std::string>> tests = {
+    std::vector<std::pair<size_t, std::string>> tests = {
         {3, "qwer"},
         {5, "1111"},
     };
